형식 : 분할 컴파일을 위한 코드 수정
변경 전
 - 코드의 분할이 안되있어서 한눈에 보이지 않는다는
 단점이 존재했다.
변경 후
 - 각각 함수의 선언 부 정의 부로 구분
  선언부 : FuncDeclare.h, 정의부 : TTTFun.c(틱택토 직접적인 관련기능)
  MIN_MAX_Alg.c (최대최소 알고리즘 구현파일)

 *추가적인 사항*
 - 현재 완벽하게 코드 분할이 되지 않았습니다.
 - 미흡한 부분은 현재 계속적으로 이것만 붙잡을 수 없기 때문에
 차후 각 팀원 코드 병합시 지속적으로 수정하겠습니다.
 - 자세하게 차후 분할할 내용은 이슈에 게시해놓겠습니다.
