+++ basedef.h
@@ -1,0 +1,25 @@
+//
+// basedef.h
+//
+// base define <basedef.h> header
+//
+
+#ifndef BASE_DEFINE
+#define BASE_DEFINE
+
+/* Define of array size */
+#define A_SIZE 3  // Array size
+
+/* Default number of return */
+#define FALSE 0
+#define TRUE 1
+
+/*Define number of game point*/
+#define WIN 10
+#define LOSE -10
+#define DRAW 0	
+
+/*Define nuber of board check*/
+enum { HUMAN, COMP, EMPTY };
+
+#endif
