+++ main.c
@@ -1,0 +1,8 @@
+#include <stdio.h>
+#include "FuncDeclare.h"
+int main(void)
+{
+	runGame();
+
+	return 0;
+}
